Improved admin header and footer. Attempted to improve user header
I've changed the layout of the header (admin) and the footer. Changed the colour scheme such that it aligns with the chosen colour pallete and made the footer more responsive by giving it interactivity in mobile view.

I've also attempted to improve Hai's code of the user header file, but was unable to do it.